Populate generation module and fix hybrid classes
