+++ src/spikeinterface/core/testing_tools.py
@@ -3,9 +3,9 @@ from __future__ import annotations
 import warnings
 
 warnings.warn(
-    "The 'testing_tools' submodule is deprecated. " "Use spikeinterface.core.generate instead",
+    "The 'testing_tools' submodule is deprecated. " "Use spikeinterface.generation instead",
     DeprecationWarning,
     stacklevel=2,
 )
 
-from .generate import *
+from ..generation.generate import *
