+++ src/spikeinterface/core/tests/test_basesorting.py
@@ -23,7 +23,7 @@ from spikeinterface.core import (
 )
 from spikeinterface.core.base import BaseExtractor
 from spikeinterface.core.testing import check_sorted_arrays_equal, check_sortings_equal
-from spikeinterface.core.generate import generate_sorting
+from spikeinterface.core import generate_sorting
 
 if hasattr(pytest, "global_test_folder"):
     cache_folder = pytest.global_test_folder / "core"
