+++ src/spikeinterface/generation/__init__.py
@@ -1,3 +1,26 @@
+from .generation_tools import (
+    generate_recording,
+    generate_sorting,
+    add_synchrony_to_sorting,
+    create_sorting_npz,
+    generate_snippets,
+    synthesize_random_firings,
+    inject_some_duplicate_units,
+    inject_some_split_units,
+    synthetize_spike_train_bad_isi,
+    generate_templates,
+    generate_recording_by_size,
+    generate_ground_truth_recording,
+    generate_channel_locations,
+    generate_unit_locations,
+    generate_single_fake_waveform,
+    generate_sorting_to_inject,
+)
+
+from .transformsorting import TransformSorting
+from .noisegeneratorrecording import NoiseGeneratorRecording, noise_generator_recording
+from .injecttemplatesrecording import InjectTemplatesRecording, inject_templates
+
 from .drift_tools import (
     move_dense_templates,
     interpolate_templates,
