+++ src/spikeinterface/preprocessing/tests/test_interpolate_bad_channels.py
@@ -5,7 +5,7 @@ import os
 import spikeinterface as si
 import spikeinterface.preprocessing as spre
 import spikeinterface.extractors as se
-from spikeinterface.core.generate import generate_recording
+from spikeinterface.generation import generate_recording
 
 try:
     import spikeglx
